+++ app/javascript/frontend/test.js
@@ -1,0 +1,1 @@
+
