+++ app/views/pens/_pens.html.erb
