+++ app/views/pens/index.html.erb
@@ -10,6 +10,7 @@
   <%= render 'deleted_pens' %>
 <% end %>
 
+<button id='get_data'>get btn</button>
 <!--<% @pens.each do |pen| %>
   <tr>
     <td><%= pen.title %></td>
